@@ -1,8 +1,8 @@
 module github.com/venjiang/wasm
 
-go 1.16
+go 1.17
 
 require (
-	github.com/mathetake/gasm v0.0.0-20210402230938-71556e50c408
-	github.com/wasmerio/wasmer-go v1.0.3 // indirect
+	github.com/tetratelabs/wazero v0.0.0-20220511011221-c50c121a48d2
+	github.com/wasmerio/wasmer-go v1.0.4
 )
